+++ python/GameOfLife/Lists.py
@@ -1,0 +1,15 @@
+
+
+class Human():
+    def __init__(self, name, birthday, gender):
+        self.name = name
+        self.birthday = birthday
+        self.gender = gender
+
+
+people_list = [
+    Human("Bob", 2006, "Robot"),
+    Human("Cath", 1980, "Female"),
+    Human("Zach", 2012, "Male"),
+    Human("Alice", 1995, "Female"),
+]
